+++ src/collections/componentSetBuilder.ts
@@ -10,8 +10,10 @@
 import * as path from 'node:path';
 import { StateAggregator, Logger, SfError, Messages } from '@salesforce/core';
 import * as fs from 'graceful-fs';
+import * as minimatch from 'minimatch';
 import { ComponentSet } from '../collections';
 import { RegistryAccess } from '../registry';
+import { FileProperties } from '../client';
 
 Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('@salesforce/source-deploy-retrieve', 'sdr');
@@ -56,6 +58,13 @@ export class ComponentSetBuilder {
   public static async build(options: ComponentSetOptions): Promise<ComponentSet> {
     const logger = Logger.childFromRoot('componentSetBuilder');
     let componentSet: ComponentSet | undefined;
+
+    /**
+     * A map used when building a ComponentSet from metadata type/name pairs
+     * key = a metadata type, e.g. `ApexClass`
+     * value = an array of metadata names, e.g. `['foo_*', 'BarClass']`
+     */
+    const mdMap = new Map<string, string[]>();
 
     const { sourcepath, manifest, metadata, packagenames, apiversion, sourceapiversion, org } = options;
     try {
@@ -103,31 +112,41 @@ export class ComponentSetBuilder {
 
         // Build a Set of metadata entries
         metadata.metadataEntries.forEach((rawEntry) => {
-          const splitEntry = rawEntry.split(':').map((entry) => entry.trim());
+          const [mdType, mdName] = rawEntry.split(':').map((entry) => entry.trim());
           // The registry will throw if it doesn't know what this type is.
-          registry.getTypeByName(splitEntry[0]);
+          registry.getTypeByName(mdType);
+
+          // Add metadata entries to a map for possible use with an org connection below.
+          const mdMapEntry = mdMap.get(mdType);
+          if (mdMapEntry) {
+            mdMapEntry.push(mdName);
+          } else {
+            mdMap.set(mdType, [mdName]);
+          }
+
           // this '.*' is a surprisingly valid way to specify a metadata, especially a DEB :sigh:
           // https://github.com/salesforcecli/plugin-deploy-retrieve/blob/main/test/nuts/digitalExperienceBundle/constants.ts#L140
           // because we're filtering from what we have locally, this won't allow you to retrieve new metadata (on the server only) using the partial wildcard
           // to do that, you'd need check the size of the CS created below, see if it's 0, and then query the org for the metadata that matches the regex
           // but building a CS from a metadata argument doesn't require an org, so we can't do that here
-          if (splitEntry[1]?.includes('*') && splitEntry[1]?.length > 1 && !splitEntry[1].includes('.*')) {
+          if (mdName?.includes('*') && mdName?.length > 1 && !mdName.includes('.*')) {
             // get all components of the type, and then filter by the regex of the fullName
             ComponentSet.fromSource({
               fsPaths: directoryPaths,
-              include: new ComponentSet([{ type: splitEntry[0], fullName: ComponentSet.WILDCARD }]),
+              include: new ComponentSet([{ type: mdType, fullName: ComponentSet.WILDCARD }]),
             })
               .getSourceComponents()
               .toArray()
-              .filter((cs) => Boolean(cs.fullName.match(new RegExp(splitEntry[1]))))
+              // using minimatch versus RegExp provides better (more expected) matching results
+              .filter((cs) => minimatch(cs.fullName, mdName))
               .map((match) => {
                 compSetFilter.add(match);
                 componentSet?.add(match);
               });
           } else {
             const entry = {
-              type: splitEntry[0],
-              fullName: splitEntry.length === 1 ? '*' : splitEntry[1],
+              type: mdType,
+              fullName: !mdName ? '*' : mdName,
             };
             // Add to the filtered ComponentSet for resolved source paths,
             // and the unfiltered ComponentSet to build the correct manifest.
@@ -147,15 +166,35 @@ export class ComponentSetBuilder {
       // Resolve metadata entries with an org connection
       if (org) {
         componentSet ??= new ComponentSet();
-        logger.debug(`Building ComponentSet from targetUsername: ${org.username}`);
+
+        let debugMsg = `Building ComponentSet from targetUsername: ${org.username}`;
+
+        // *** Default Filter ***
+        // exclude components based on the results of componentFilter function
+        // components with namespacePrefix where org.exclude includes manageableState (to exclude managed packages)
+        // components with namespacePrefix where manageableState equals undefined (to exclude components e.g. InstalledPackage)
+        // components where org.exclude includes manageableState (to exclude packages without namespacePrefix e.g. unlocked packages)
+        let componentFilter = (component: Partial<FileProperties>): boolean =>
+          !component?.manageableState || !org.exclude?.includes(component.manageableState);
+
+        if (metadata) {
+          debugMsg += ` filtered by metadata: ${metadata.metadataEntries.toString()}`;
+
+          componentFilter = (component: Partial<FileProperties>): boolean => {
+            if (component.type && component.fullName) {
+              const mdMapEntry = mdMap.get(component.type);
+              // using minimatch versus RegExp provides better (more expected) matching results
+              return !!mdMapEntry && mdMapEntry.some((mdName) => minimatch(component.fullName as string, mdName));
+            }
+            return false;
+          };
+        }
+
+        logger.debug(debugMsg);
         const fromConnection = await ComponentSet.fromConnection({
           usernameOrConnection: (await StateAggregator.getInstance()).aliases.getUsername(org.username) ?? org.username,
-          // exclude components based on the results of componentFilter function
-          // components with namespacePrefix where org.exclude includes manageableState (to exclude managed packages)
-          // components with namespacePrefix where manageableState equals undefined (to exclude components e.g. InstalledPackage)
-          // components where org.exclude includes manageableState (to exclude packages without namespacePrefix e.g. unlocked packages)
-          componentFilter: (component): boolean =>
-            !component?.manageableState || !org.exclude?.includes(component.manageableState),
+          componentFilter,
+          metadataTypes: mdMap.size ? Array.from(mdMap.keys()) : undefined,
         });
 
         for (const comp of fromConnection) {