+++ src/resolve/forceIgnore.ts
@@ -21,22 +21,19 @@ export class ForceIgnore {
 
   public constructor(forceIgnorePath = '') {
     try {
-      let contents = readFileSync(forceIgnorePath, 'utf-8');
+      const contents = readFileSync(forceIgnorePath, 'utf-8');
       // check if file `.forceignore` exists
       if (contents !== undefined) {
         // check for windows style separators (\) and warn
         if (contents.includes('\\')) {
-          const lifecycle = Lifecycle.getInstance();
-          // cannot await a method in a constructor
-          void lifecycle.emitWarning(
-            'Your .forceignore file incorrectly uses the backslash ("\\") as a folder separator; it should use the slash ("/") instead. We currently accept both separators, but we plan to stop supporting the backslash soon.'
+          // void because you cannot await a method in a constructor
+          void Lifecycle.getInstance().emitWarning(
+            'Your .forceignore file incorrectly uses the backslash ("\\") as a folder separator; it should use the slash ("/") instead. The ignore rules will not work as expected until you fix this.'
           );
-          // TODO: change this in v56 to only emit warning but NOT fix file
-          contents = contents.replace(/\\/g, '/');
         }
-
         // add the default ignore paths, and then parse the .forceignore file
-        this.parser = ignore().add(`${contents}\n${this.DEFAULT_IGNORE.join('\n')}`);
+        this.parser = ignore().add(`${this.DEFAULT_IGNORE.join('\n')}\n${contents}`);
+
         this.forceIgnoreDirectory = dirname(forceIgnorePath);
       }
     } catch (e) {